+++ idp/authmiddleware/authmiddleware_test.go
@@ -1,10 +1,11 @@
-package idp_test
+package authmiddleware_test
 
 import (
 	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
+	"github.com/d-velop/dvelop-sdk-go/idp/authmiddleware"
 	"net/http"
 	"net/http/httptest"
 	"reflect"
@@ -12,7 +13,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/d-velop/dvelop-sdk-go/idp"
 	"github.com/d-velop/dvelop-sdk-go/idp/scim"
 )
 
@@ -25,7 +25,7 @@ func TestGetRequestWithFalseAuthorizationType_RedirectsToIdp(t *testing.T) {
 	req.Header.Set("Authorization", "Basic adadbk")
 	handlerSpy := handlerSpy{}
 
-	idp.Authenticate(nil, nil, false, log, log)(&handlerSpy).ServeHTTP(responseSpy, req)
+	authmiddleware.Authenticate(nil, nil, false, log, log)(&handlerSpy).ServeHTTP(responseSpy, req)
 
 	if err := responseSpy.assertStatusCodeIs(http.StatusFound); err != nil {
 		t.Error(err)
@@ -47,7 +47,7 @@ func TestGetRequestWithFalseCookie_RedirectsToIdp(t *testing.T) {
 	req.AddCookie(&http.Cookie{Name: "AnyCookie", Value: "adadbk"})
 	handlerSpy := handlerSpy{}
 
-	idp.Authenticate(nil, nil, false, log, log)(&handlerSpy).ServeHTTP(responseSpy, req)
+	authmiddleware.Authenticate(nil, nil, false, log, log)(&handlerSpy).ServeHTTP(responseSpy, req)
 
 	if err := responseSpy.assertStatusCodeIs(http.StatusFound); err != nil {
 		t.Error(err)
@@ -67,7 +67,7 @@ func TestHeadRequestWithoutAuthorizationInfos_RedirectsToIdp(t *testing.T) {
 	}
 	responseSpy := responseSpy{httptest.NewRecorder()}
 	handlerSpy := &handlerSpy{}
-	idp.Authenticate(nil, nil, false, log, log)(handlerSpy).ServeHTTP(responseSpy, req)
+	authmiddleware.Authenticate(nil, nil, false, log, log)(handlerSpy).ServeHTTP(responseSpy, req)
 
 	if err := responseSpy.assertStatusCodeIs(http.StatusFound); err != nil {
 		t.Error(err)
@@ -87,7 +87,7 @@ func TestPostRequestWithoutAuthorizationInfos_ReturnsStatus401(t *testing.T) {
 	}
 	responseSpy := responseSpy{httptest.NewRecorder()}
 	handlerSpy := &handlerSpy{}
-	idp.Authenticate(nil, nil, false, log, log)(handlerSpy).ServeHTTP(responseSpy, req)
+	authmiddleware.Authenticate(nil, nil, false, log, log)(handlerSpy).ServeHTTP(responseSpy, req)
 
 	if err := responseSpy.assertStatusCodeIs(http.StatusUnauthorized); err != nil {
 		t.Error(err)
@@ -107,7 +107,7 @@ func TestPutRequestWithoutAuthorizationInfos_ReturnsStatus401(t *testing.T) {
 	}
 	responseSpy := responseSpy{httptest.NewRecorder()}
 	handlerSpy := &handlerSpy{}
-	idp.Authenticate(nil, nil, false, log, log)(handlerSpy).ServeHTTP(responseSpy, req)
+	authmiddleware.Authenticate(nil, nil, false, log, log)(handlerSpy).ServeHTTP(responseSpy, req)
 
 	if err := responseSpy.assertStatusCodeIs(http.StatusUnauthorized); err != nil {
 		t.Error(err)
@@ -127,7 +127,7 @@ func TestDeleteRequestWithoutAuthorizationInfos_ReturnsStatus401(t *testing.T) {
 	}
 	responseSpy := responseSpy{httptest.NewRecorder()}
 	handlerSpy := &handlerSpy{}
-	idp.Authenticate(nil, nil, false, log, log)(handlerSpy).ServeHTTP(responseSpy, req)
+	authmiddleware.Authenticate(nil, nil, false, log, log)(handlerSpy).ServeHTTP(responseSpy, req)
 
 	if err := responseSpy.assertStatusCodeIs(http.StatusUnauthorized); err != nil {
 		t.Error(err)
@@ -147,7 +147,7 @@ func TestPatchRequestWithoutAuthorizationInfos_ReturnsStatus401(t *testing.T) {
 	}
 	responseSpy := responseSpy{httptest.NewRecorder()}
 	handlerSpy := &handlerSpy{}
-	idp.Authenticate(nil, nil, false, log, log)(handlerSpy).ServeHTTP(responseSpy, req)
+	authmiddleware.Authenticate(nil, nil, false, log, log)(handlerSpy).ServeHTTP(responseSpy, req)
 
 	if err := responseSpy.assertStatusCodeIs(http.StatusUnauthorized); err != nil {
 		t.Error(err)
@@ -176,7 +176,7 @@ func TestRequestWithBearerAuthorization_PopulatesContextWithPrincipalAndAuthsess
 	idpStub := newIdpStub(map[string]scim.Principal{authSessionId: principal}, nil)
 	defer idpStub.Close()
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
 
 	if err := handlerSpy.assertAuthSessionIdIs(authSessionId); err != nil {
 		t.Error(err)
@@ -198,7 +198,7 @@ func TestRequestWithLowerCaseBearerAuthorization_PopulatesContextWithPrincipalAn
 	idpStub := newIdpStub(map[string]scim.Principal{authSessionId: principal}, nil)
 	defer idpStub.Close()
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
 
 	if err := handlerSpy.assertAuthSessionIdIs(authSessionId); err != nil {
 		t.Error(err)
@@ -221,7 +221,7 @@ func TestRequestWithAuthSessionIdCookie_PopulatesContextWithPrincipalAndAuthsess
 	idpStub := newIdpStub(map[string]scim.Principal{authSessionId: principal}, nil)
 	defer idpStub.Close()
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
 
 	if err := handlerSpy.assertAuthSessionIdIs(authSessionId); err != nil {
 		t.Error(err)
@@ -245,7 +245,7 @@ func TestRequestWithBadUrlEncodedAuthSessionIdCookie_ReturnsStatus500(t *testing
 	defer idpStub.Close()
 	spy := responseSpy{httptest.NewRecorder()}
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
 
 	if err := spy.assertStatusCodeIs(http.StatusInternalServerError); err != nil {
 		t.Error(err)
@@ -269,7 +269,7 @@ func TestRequestWithBearerTokenAndCookie_PopulatesContextUsingBearerToken(t *tes
 	idpStub := newIdpStub(map[string]scim.Principal{authSessionId: principal}, nil)
 	defer idpStub.Close()
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
 
 	if err := handlerSpy.assertAuthSessionIdIs(authSessionId); err != nil {
 		t.Error(err)
@@ -295,7 +295,7 @@ func TestRequestWithBadTokenAndExternalValidationIsNotAllowed_RedirectsToIdp(t *
 	defer idpStub.Close()
 	spy := responseSpy{httptest.NewRecorder()}
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
 
 	if err := spy.assertStatusCodeIs(http.StatusFound); err != nil {
 		t.Error(err)
@@ -324,7 +324,7 @@ func TestRequestWithBadTokenAndExternalValidationIsAllowed_RedirectsToIdp(t *tes
 	defer idpStub.Close()
 	spy := responseSpy{httptest.NewRecorder()}
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), true, log, log)(&handlerSpy).ServeHTTP(spy, req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), true, log, log)(&handlerSpy).ServeHTTP(spy, req)
 
 	if err := spy.assertStatusCodeIs(http.StatusFound); err != nil {
 		t.Error(err)
@@ -350,7 +350,7 @@ func TestGetSystemBaseUriFromCtxReturnsError_ReturnsStatus500(t *testing.T) {
 	defer idpStub.Close()
 	spy := responseSpy{httptest.NewRecorder()}
 
-	idp.Authenticate(func(ctx context.Context) (string, error) { return "", errors.New("any error") }, returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
+	authmiddleware.Authenticate(func(ctx context.Context) (string, error) { return "", errors.New("any error") }, returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
 
 	if err := spy.assertStatusCodeIs(http.StatusInternalServerError); err != nil {
 		t.Error(err)
@@ -373,7 +373,7 @@ func TestGetTenantIdFromCtxReturnsError_ReturnsStatus500(t *testing.T) {
 	defer idpStub.Close()
 	spy := responseSpy{httptest.NewRecorder()}
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), func(ctx context.Context) (string, error) { return "", errors.New("any error") }, false, log, log)(&handlerSpy).ServeHTTP(spy, req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), func(ctx context.Context) (string, error) { return "", errors.New("any error") }, false, log, log)(&handlerSpy).ServeHTTP(spy, req)
 
 	if err := spy.assertStatusCodeIs(http.StatusInternalServerError); err != nil {
 		t.Error(err)
@@ -398,7 +398,7 @@ func TestIdPReturnsStatus500_ReturnsStatus500(t *testing.T) {
 	defer idpStub.Close()
 	spy := responseSpy{httptest.NewRecorder()}
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
 
 	if err := spy.assertStatusCodeIs(http.StatusInternalServerError); err != nil {
 		t.Error(err)
@@ -424,7 +424,7 @@ func TestIdPReturnsMalformedJson_ReturnsStatus500(t *testing.T) {
 	defer idpStub.Close()
 	spy := responseSpy{httptest.NewRecorder()}
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
 
 	if err := spy.assertStatusCodeIs(http.StatusInternalServerError); err != nil {
 		t.Error(err)
@@ -450,7 +450,7 @@ func TestIdPReturnsPrincipalWithEmptyId_ReturnsStatus500(t *testing.T) {
 	defer idpStub.Close()
 	spy := responseSpy{httptest.NewRecorder()}
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
 
 	if err := spy.assertStatusCodeIs(http.StatusInternalServerError); err != nil {
 		t.Error(err)
@@ -479,9 +479,9 @@ func TestUserIsCachedAndCacheEntryIsNotExpired_ReturnsCachedEntry(t *testing.T) 
 	}))
 	defer idpStub.Close()
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
 	time.Sleep(1 * time.Nanosecond)
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
 
 	if idpCalled != 1 {
 		t.Errorf("IdP has been called %v times but expected %v times", idpCalled, 1)
@@ -510,9 +510,9 @@ func TestUserIsCachedButCacheEntryIsExpired_CallsIdp(t *testing.T) {
 	}))
 	defer idpStub.Close()
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
 	time.Sleep(1 * time.Second)
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
 
 	if idpCalled != 2 {
 		t.Errorf("IdP has been called %v times but expected %v times", idpCalled, 2)
@@ -534,7 +534,7 @@ func TestUserIsCachedForDifferentTenant_CallsIdp(t *testing.T) {
 	idpStub1 := newIdpStub(map[string]scim.Principal{authSessionId: principalT1}, nil)
 	defer idpStub1.Close()
 	handlerSpy1 := handlerSpy{}
-	idp.Authenticate(returnFromCtx(idpStub1.URL), returnFromCtx("1"), false, log, log)(&handlerSpy1).ServeHTTP(httptest.NewRecorder(), reqTenant1)
+	authmiddleware.Authenticate(returnFromCtx(idpStub1.URL), returnFromCtx("1"), false, log, log)(&handlerSpy1).ServeHTTP(httptest.NewRecorder(), reqTenant1)
 
 	if err := handlerSpy1.assertPrincipalIs(principalT1); err != nil {
 		t.Error(err)
@@ -549,7 +549,7 @@ func TestUserIsCachedForDifferentTenant_CallsIdp(t *testing.T) {
 	idpStub2 := newIdpStub(map[string]scim.Principal{authSessionId: principalT2}, nil)
 	defer idpStub2.Close()
 	handlerSpy2 := handlerSpy{}
-	idp.Authenticate(returnFromCtx(idpStub2.URL), returnFromCtx("2"), false, log, log)(&handlerSpy2).ServeHTTP(httptest.NewRecorder(), reqTenant2)
+	authmiddleware.Authenticate(returnFromCtx(idpStub2.URL), returnFromCtx("2"), false, log, log)(&handlerSpy2).ServeHTTP(httptest.NewRecorder(), reqTenant2)
 
 	if err := handlerSpy2.assertPrincipalIs(principalT2); err != nil {
 		t.Error(err)
@@ -568,7 +568,7 @@ func TestRequestAsExternalUserAndExternalUserValidationIsNotAllowed_ReturnsStatu
 	defer idpStub.Close()
 	spy := responseSpy{httptest.NewRecorder()}
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(spy, req)
 	if err := spy.assertStatusCodeIs(http.StatusForbidden); err != nil {
 		t.Error(err)
 	}
@@ -589,7 +589,7 @@ func TestRequestAsExternalUserAndExternalValidationIsAllowed_PopulatesContextWit
 	idpStub := newIdpStub(nil, map[string]scim.Principal{authSessionId: principal})
 	defer idpStub.Close()
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), true, log, log)(handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), true, log, log)(handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
 
 	if err := handlerSpy.assertAuthSessionIdIs(authSessionId); err != nil {
 		t.Error(err)
@@ -611,7 +611,7 @@ func TestRequestAsInternalUserAndExternalValidationIsAllowed_PopulatesContextWit
 	idpStub := newIdpStub(map[string]scim.Principal{authSessionId: principal}, nil)
 	defer idpStub.Close()
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), true, log, log)(handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), true, log, log)(handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
 
 	if err := handlerSpy.assertAuthSessionIdIs(authSessionId); err != nil {
 		t.Error(err)
@@ -639,8 +639,8 @@ func TestIdpSendsNoCacheHeader_CallsIdp(t *testing.T) {
 	}))
 	defer idpStub.Close()
 
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
-	idp.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
+	authmiddleware.Authenticate(returnFromCtx(idpStub.URL), returnFromCtx("1"), false, log, log)(&handlerSpy).ServeHTTP(httptest.NewRecorder(), req)
 
 	if idpCalled != 2 {
 		t.Errorf("IdP has been called %v times but expected %v times", idpCalled, 2)
@@ -690,7 +690,7 @@ func TestGetRequestWithoutAuthorizationInfosWithAcceptHeader(t *testing.T) {
 			responseSpy := responseSpy{httptest.NewRecorder()}
 
 			handlerSpy := &handlerSpy{}
-			idp.Authenticate(nil, nil, false, log, log)(handlerSpy).ServeHTTP(responseSpy, req)
+			authmiddleware.Authenticate(nil, nil, false, log, log)(handlerSpy).ServeHTTP(responseSpy, req)
 
 			if handlerSpy.hasBeenCalled {
 				t.Error("inner handler should not have been called")
@@ -750,14 +750,14 @@ func newIdpStub(principals map[string]scim.Principal, externalPrincipals map[str
 
 type handlerSpy struct {
 	authSessionId string
-	prinicpal     scim.Principal
+	principal     scim.Principal
 	hasBeenCalled bool
 }
 
 func (spy *handlerSpy) ServeHTTP(rw http.ResponseWriter, r *http.Request) {
 	spy.hasBeenCalled = true
-	spy.authSessionId, _ = idp.AuthSessionIdFromCtx(r.Context())
-	spy.prinicpal, _ = idp.PrincipalFromCtx(r.Context())
+	spy.authSessionId, _ = authmiddleware.AuthSessionIdFromCtx(r.Context())
+	spy.principal, _ = authmiddleware.PrincipalFromCtx(r.Context())
 }
 
 func (spy *handlerSpy) assertAuthSessionIdIs(expectedAuthSessionID string) error {
@@ -768,8 +768,8 @@ func (spy *handlerSpy) assertAuthSessionIdIs(expectedAuthSessionID string) error
 }
 
 func (spy *handlerSpy) assertPrincipalIs(expectedPrincipal scim.Principal) error {
-	if !reflect.DeepEqual(spy.prinicpal, expectedPrincipal) {
-		return fmt.Errorf("handler set wrong principal on context: got \n %v want\n %v", spy.prinicpal, expectedPrincipal)
+	if !reflect.DeepEqual(spy.principal, expectedPrincipal) {
+		return fmt.Errorf("handler set wrong principal on context: got \n %v want\n %v", spy.principal, expectedPrincipal)
 	}
 	return nil
 }
